// src/apps/experimental/routes/user/display/types.ts
export interface DisplaySettingsValues {
    customCss: string;
    dashboardTheme: string;
    dateTimeLocale: string;
    disableCustomCss: boolean;
    displayMissingEpisodes: boolean;
    enableBlurHash: boolean;
    enableFasterAnimation: boolean;
    enableItemDetailsBanner: boolean;
    enableLibraryBackdrops: boolean;
    enableLibraryThemeSongs: boolean;
    enableLibraryThemeVideos: boolean;
    enableRewatchingInNextUp: boolean;
    episodeImagesInNextUp: boolean;
    language: string;
    layout: string;
    libraryPageSize: number;
    maxDaysForNextUp: number;
    screensaver: string;
    screensaverInterval: number;
    theme: string;
}